I think you can get jobs now

+++ Jobber.Sdk/Models/Financials/Invoice.cs
@@ -15,13 +15,13 @@ namespace Jobber.Sdk.Models.Financials
         public string Message { get; set; }
 
         [JsonProperty("total")]
-        public int Total { get; set; }
+        public double Total { get; set; }
 
         [JsonProperty("draft")]
         public bool Draft { get; set; }
 
         [JsonProperty("invoice_net")]
-        public int InvoiceNet { get; set; }
+        public double InvoiceNet { get; set; }
 
         [JsonProperty("bad_debt")]
         public bool BadDebt { get; set; }
